rename InstrumentedChild to CachedChild
