+++ NDChild.py
@@ -1,3 +1,7 @@
+import re
+from utils import progress_bar
+
+
 def format_val(val):
     if isinstance(val, float):
         return round(val, 2)
@@ -374,3 +378,97 @@ class NDChild(object):
             self.grammar[parameter] -= self.conservativerate * self.grammar[parameter]
         elif direction == 1:
             self.grammar[parameter] += self.conservativerate * (1 - self.grammar[parameter])
+
+
+class TriggerCacher:
+    def __init__(self):
+        self.current_trigger = None
+        self.current_sentence = None
+
+    def trigger_name(self, method):
+        name = re.sub('Etrigger', '', method.__name__)
+        name = name.upper() if name.islower() else name
+        return name
+
+    def set_trigger(self, source_trigger, sentence, param, direction, conservative: bool):
+        trigs = sentence.triggers
+        if source_trigger not in trigs:
+            trigs[source_trigger] = [(param, direction, conservative)]
+        else:
+            trigs[source_trigger].append((param, direction, conservative))
+
+    def adjustweight(self, param, direction):
+        self.set_trigger(self.current_trigger, self.current_sentence, param, direction, False)
+
+    def adjustweightConservatively(self, param, direction):
+        self.set_trigger(self.current_trigger, self.current_sentence, param, direction, True)
+
+    def consume_sentence(self, s):
+        child = NDChild(None, None, None)
+        child.adjustweight = self.adjustweight
+        child.adjustweightConservatively = self.adjustweightConservatively
+        for trigger in child.trigger_methods:
+            self.current_trigger = self.trigger_name(trigger)
+            self.current_sentence = s
+            trigger(s)
+        self.current_trigger = None
+        self.current_sentence = None
+
+
+class CachedChild(NDChild):
+    """An implementation of NDChild where all triggers which are pure functions,
+    (whose behavior depends solely on the input sentence, and not the state of
+    the grammar) are cached once for every sentence in the domain and then
+    simply looked up at runtime.
+
+    """
+    cacher = TriggerCacher()
+
+    @classmethod
+    def precompute_domain(cls, domain):
+        for sentence in progress_bar(domain.sentences.values(),
+                                     desc='precomputing triggers'):
+            cls.precompute_sentence(sentence)
+
+    @classmethod
+    def precompute_sentence(cls, sentence):
+        cls.cacher.consume_sentence(sentence)
+
+    def handleTrigger(self, sentence, etrigger):
+        if etrigger not in sentence.triggers:
+            return
+        for param, direction, conservative in sentence.triggers[etrigger]:
+            if conservative:
+                self.adjustweightConservatively(param, direction)
+            else:
+                self.adjustweight(param, direction)
+
+    def QInvEtrigger(self, s):
+        self.handleTrigger(s, 'QInv')
+
+    def VtoIEtrigger(self, s):
+        self.handleTrigger(s, 'VtoI')
+
+    def hcpEtrigger(self, s):
+        self.handleTrigger(s, 'HCP')
+
+    def hipEtrigger(self, s):
+        self.handleTrigger(s, 'HIP')
+
+    def nsEtrigger(self, s):
+        self.handleTrigger(s, 'NS')
+
+    def ntEtrigger(self, s):
+        self.handleTrigger(s, 'NT')
+
+    def piEtrigger(self, s):
+        self.handleTrigger(s, 'PI')
+
+    def spEtrigger(self, s):
+        self.handleTrigger(s, 'SP')
+
+    def tmEtrigger(self, s):
+        self.handleTrigger(s, 'TM')
+
+    def whmEtrigger(self, s):
+        self.handleTrigger(s, 'WHM')
